+++ plugin/__init__.py
@@ -1,4 +1,7 @@
 """Pinger worker plugin package."""
 
-__version_info__: tuple[int, int, int] = (0, 0, 1)
+from typing import Tuple
+
+
+__version_info__: Tuple[int, int, int] = (0, 0, 2)
 __version__: str = ".".join(map(str, __version_info__))